pyproject.toml
[tool.poetry]
name = "HQ-Animate"
version = "0.1.0"
license = "MIT"
description = "Create high quality APNG and WebP animations from a sequence of images."
authors = ["Ethan Chappel <EthanChappel@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
pillow = "^11.3.0"
numpy = "^2.3.2"
astropy = "^7.1.0"
pyside6-essentials = "^6.9.1"

[tool.poetry.group.dev.dependencies]
cx_freeze = "^8.3.0"
pip-licenses = "^5.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
